src/domain/forum/application/use-cases/comment-on-answer.spec.ts
import { describe, expect, it } from 'vitest'
import { UniqueEntityID } from '@/core/entities/unique-entity-id'
import { CommentOnAnswerUseCase } from './comment-on-answer'
import { InMemoryAnswerCommentsRepository } from 'test/repositories/in-memory-answers-comments-repository'
import { InMemoryAnswersRepository } from 'test/repositories/in-memory-answers-repository'
import { makeAnswer } from 'test/factories/make-answer'

let sut: CommentOnAnswerUseCase
let inMemoryAnswerCommentsRepository: InMemoryAnswerCommentsRepository
let inMemoryAnswerRepository: InMemoryAnswersRepository

describe('Comment On Answer Use Case', () => {
  beforeEach(() => {
    inMemoryAnswerRepository = new InMemoryAnswersRepository()
    inMemoryAnswerCommentsRepository = new InMemoryAnswerCommentsRepository()
    sut = new CommentOnAnswerUseCase(
      inMemoryAnswerRepository,
      inMemoryAnswerCommentsRepository,
    )
  })

  it('should be able to comment on answer', async () => {
    const newAnswer = makeAnswer()

    await inMemoryAnswerRepository.create(newAnswer)

    const { answerComment } = await sut.execute({
      authorId: 'author-1',
      answerId: newAnswer.id.toString(),
      content: 'comment on answer',
    })

    expect(answerComment.id).toBeInstanceOf(UniqueEntityID)
    expect(answerComment.content).toEqual('comment on answer')
    expect(inMemoryAnswerCommentsRepository.items[0].content).toEqual(
      'comment on answer',
    )
  })
})
